moved most block files to be common- some of them will be used by the item generator, others will be for the diorama generator.
