+++ src/generators/_common/plugins/texturePicker/face.ts
@@ -6,13 +6,8 @@ import {
   type Generator,
   type Region,
 } from "@genroot/builder/modules/generator";
-import { currentBlockTextureId } from "./constants";
-import {
-  type SelectedTextureWithBlend,
-  decodeSelectedTextureWithBlend,
-  decodeSelectedTextureWithBlendArray,
-  encodeSelectedTextureWithBlendArray,
-} from "./selectedTextureWithBlend";
+import { decodeSelectedTextureWithBlend, decodeSelectedTextureWithBlendArray, encodeSelectedTextureWithBlendArray, SelectedTextureWithBlend } from "./selectedTextureWithBlend";
+
 
 export function defineInputRegion(
   generator: Generator,
@@ -21,7 +16,7 @@ export function defineInputRegion(
 ) {
   generator.defineRegionInput(region, () => {
     const selectedTextureJson = generator.getStringInputValue(
-      currentBlockTextureId
+      "CurrentBlockTexture"
     );
 
     const selectedTexture = selectedTextureJson
