+++ src/generators/_common/plugins/texturePicker/selectedTextureWithBlend.ts
