+++ src/generators/_common/plugins/texturePicker/tintSelector.tsx
