+++ src/generators/_common/plugins/texturePicker/tints.ts
